minor: rename srcs
