+++ nivel-gui/nivel-gui.h
