+++ nivel-gui/tad_nivel.c
@@ -1,5 +1,5 @@
 
-#include "tad_items.h"
+#include "tad_nivel.h"
 
 #include <stdlib.h>
 #include <stdio.h>
